update: lock feature completed
